refactor(Moved a few things around, updated comments and added a method to gen trip lat/lon):
